- Factor out shared `hasDefinedTypeReturnTag` and `hasReturnValue` code, applying to `requireReturnsCheck`
